gitserver: Move some VCSSyncer code around
Moved each implemenation into their own files and reordered some methods
so that all the public ones are first.
