+++ cmd/gitserver/server/vcs_syncer_git.go
@@ -1,0 +1,103 @@
+package server
+
+import (
+	"context"
+	"os"
+	"os/exec"
+
+	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/vcs"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
+)
+
+// GitRepoSyncer is a syncer for Git repositories.
+type GitRepoSyncer struct{}
+
+func (s *GitRepoSyncer) Type() string {
+	return "git"
+}
+
+// IsCloneable checks to see if the Git remote URL is cloneable.
+func (s *GitRepoSyncer) IsCloneable(ctx context.Context, remoteURL *vcs.URL) error {
+	if isAlwaysCloningTest(api.RepoName(remoteURL.String())) {
+		return nil
+	}
+	if testGitRepoExists != nil {
+		return testGitRepoExists(ctx, remoteURL)
+	}
+
+	args := []string{"ls-remote", remoteURL.String(), "HEAD"}
+	ctx, cancel := context.WithTimeout(ctx, shortGitCommandTimeout(args))
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "git", args...)
+	out, err := runWithRemoteOpts(ctx, cmd, nil)
+	if err != nil {
+		if ctxerr := ctx.Err(); ctxerr != nil {
+			err = ctxerr
+		}
+		if len(out) > 0 {
+			err = errors.Errorf("%s (output follows)\n\n%s", err, out)
+		}
+		return err
+	}
+	return nil
+}
+
+// CloneCommand returns the command to be executed for cloning a Git repository.
+func (s *GitRepoSyncer) CloneCommand(ctx context.Context, remoteURL *vcs.URL, tmpPath string) (cmd *exec.Cmd, err error) {
+	if err := os.MkdirAll(tmpPath, os.ModePerm); err != nil {
+		return nil, errors.Wrapf(err, "clone failed to create tmp dir")
+	}
+
+	cmd = exec.CommandContext(ctx, "git", "init", "--bare", ".")
+	cmd.Dir = tmpPath
+	if err := cmd.Run(); err != nil {
+		return nil, errors.Wrapf(err, "clone setup failed")
+	}
+
+	cmd, _ = s.fetchCommand(ctx, remoteURL)
+	cmd.Dir = tmpPath
+	return cmd, nil
+}
+
+// Fetch tries to fetch updates of a Git repository.
+func (s *GitRepoSyncer) Fetch(ctx context.Context, remoteURL *vcs.URL, dir GitDir) error {
+	cmd, configRemoteOpts := s.fetchCommand(ctx, remoteURL)
+	dir.Set(cmd)
+	if output, err := runWith(ctx, cmd, configRemoteOpts, nil); err != nil {
+		return errors.Wrapf(err, "failed to update with output %q", newURLRedactor(remoteURL).redact(string(output)))
+	}
+	return nil
+}
+
+// RemoteShowCommand returns the command to be executed for showing remote of a Git repository.
+func (s *GitRepoSyncer) RemoteShowCommand(ctx context.Context, remoteURL *vcs.URL) (cmd *exec.Cmd, err error) {
+	return exec.CommandContext(ctx, "git", "remote", "show", remoteURL.String()), nil
+}
+
+func (s *GitRepoSyncer) fetchCommand(ctx context.Context, remoteURL *vcs.URL) (cmd *exec.Cmd, configRemoteOpts bool) {
+	configRemoteOpts = true
+	if customCmd := customFetchCmd(ctx, remoteURL); customCmd != nil {
+		cmd = customCmd
+		configRemoteOpts = false
+	} else if useRefspecOverrides() {
+		cmd = refspecOverridesFetchCmd(ctx, remoteURL)
+	} else {
+		cmd = exec.CommandContext(ctx, "git", "fetch",
+			"--progress", "--prune", remoteURL.String(),
+			// Normal git refs
+			"+refs/heads/*:refs/heads/*", "+refs/tags/*:refs/tags/*",
+			// GitHub pull requests
+			"+refs/pull/*:refs/pull/*",
+			// GitLab merge requests
+			"+refs/merge-requests/*:refs/merge-requests/*",
+			// Bitbucket pull requests
+			"+refs/pull-requests/*:refs/pull-requests/*",
+			// Gerrit changesets
+			"+refs/changes/*:refs/changes/*",
+			// Possibly deprecated refs for sourcegraph zap experiment?
+			"+refs/sourcegraph/*:refs/sourcegraph/*")
+	}
+	return cmd, configRemoteOpts
+}
